create a useFetch custom hook to be cleaner, still using graphql
