+++ henriqueApp/app/users.tsx
@@ -1,72 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { View, Text, FlatList, StyleSheet, ActivityIndicator } from 'react-native';
-import AsyncStorage from '@react-native-async-storage/async-storage';
+import { useFetchUsers } from '../hooks/useFetchUsers';
 
-type User = {
-  id: string;
-  name: string;
-  email: string;
-};
+export default function UsersList() {
+  const { users, loading, error } = useFetchUsers();
 
-const USERS_QUERY = `
-  query {
-    users {
-      nodes {
-        id
-        name
-        email
-      }
-    }
-  }
-`;
-
-const API_URL = 'https://template-onboarding-node-sjz6wnaoia-uc.a.run.app/graphql';
-
-export default function DetailsScreen() {
-  const [users, setUsers] = useState<User[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    const fetchUsers = async () => {
-      try {
-        const token = await AsyncStorage.getItem('authToken');
-        
-        if (!token) {
-          setError('Authorization token not found');
-          setLoading(false);
-          return;
-        }
-
-        const response = await fetch(API_URL, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            'Authorization': token,
-          },
-          body: JSON.stringify({ query: USERS_QUERY }),
-        });
-
-        const result = await response.json();
-
-        if (response.ok) {
-          setUsers(result.data.users.nodes);
-          console.log(response)
-        } else {
-          setError(result.errors?.[0]?.message || 'Failed to fetch users');
-        }
-      } catch (error) {
-        console.error("Failed to fetch users:", error);
-        setError("Connection error. Try again.");
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchUsers();
-  }, []);
-
-  const renderItem = ({ item }: { item: User }) => (
+  const renderItem = ({ item }: { item: { name: string; email: string } }) => (
     <View style={styles.userItem}>
       <Text style={styles.userName}>{item.name}</Text>
       <Text style={styles.userEmail}>{item.email}</Text>
